Added a section "How it works" in the spectate screen and also a tab.
Removed the 'Practice' tab
