+++ src/languages/en_GB.json
@@ -126,7 +126,7 @@
     "hostOrJoin": "Host or Join",
     "spectate": "Spectate",
     "practice": "Practice",
-    "help": "Help"
+    "howItWorks": "How it works"
   },
   "instructions": {
     "title": "Instructions",
